ansible: Use inventory dir and remove duplicate vars
Some ansible inventory variables are not needed per node but are
common for the whole setup. So move them to the hardware_base class.
Also create an inventory directory instead of a single file. That way,
it's easier to make adjustments just by adding new files (which will
be merged by Ansible itself).

Signed-off-by: Thomas Bechtold <tbechtold@suse.com>