Add caasp support
Caasp deployment is making use of skuba deployment tool.
This requires adding support for an ssh-agent
Ideally would need a loadbalancer too but this is not
implemented yet due to libcloud provider not supporting
LB on openstack.
Use "raw" (i.e yaml) ansible playbook. The code is written
so that adding another distro support should be pretty easy.

Renames kubernetes folder to k8s which is shorter
and widely used