tests/lib/hardware: Add libvirt hardware backend
There is already a OpenStack backend for providing the hardware. But
for local testing, a libvirt backend might be faster. So adding one
here.

The backend can be used by exporting "CLOUD_PROVIDER=LIBVIRT". Then
there needs to be a qcow2 image that includes cloud-init.
That image path needs to be exported as
ROOKCHECK_PROVIDER_LIBVIRT_IMAGE so rookcheck is aware of the image
path.
The provided image is not modified in any way. Instead, a qcow2
snapshot file is created (currently in the same dir where the image
lives) and that snapshot image will be used.
To get the ssh key into the image, cloud-init "nocloud" with a config
seed image is used.

Note: libvirt-python is currently not installed in the tox venv. This
needs to be done manually.

Signed-off-by: Thomas Bechtold <tbechtold@suse.com>