hardware/hardware_base: Add masters and workers properties
Mark the method get_node_by_role() as private given that it should
only be used internally. Then also drop the get_masters() and
get_workers() functions and just add properties for master and
worker.
That's what is used in the end and in the caasp implementation, there
is no longer a need to call get_masters() and get_workers().

Signed-off-by: Thomas Bechtold <tbechtold@suse.com>