kubernetes/kubernetes_base: Introduce a bootstrap() abstract method
For testing, we need a way to

1) bootstrap a cluster so a single master is available
2) add master/worker nodes to the cluster
3) remove master/worker nodes from the cluster

Currently, the structure is that we can only do a whole setup but
there is no way to do step 2) or 3).
Adding a bootstrap method (step 1)) is a first step to structure the
kubernetes installation a bit differently to be able to support the
usecases later.

Signed-off-by: Thomas Bechtold <tbechtold@suse.com>