Fix logging of ssh-key commands
For some reason, ssh-add outputs to stderr which
will be logged as a warning. It's not really
dangerous because we're creating and destroying
our own agent, so we'll suppress the messages.

Additionally ssh-keygen -R will give a warning if the IP
does not exist. We don't need to log a warning, so suppress
this too.