Disable logger means don't output anything
If you supply None to stdout and stderr in subprocess.Popen, the output
is not redirect and will still go to sys.stdout/sys.stderr. For the helper
function, we either want to log those to the logger, or ignore them
altogether when they are too verbose. Therefore, if disable_logger is set
to True, then redirect stdout and stderr to /dev/null.