kubernetes/vanilla: Start using standard ansible yml files
Instead of defining ansible tasks in python code (the API for that
is *not* stable), start using standard ansible roles & tasks.
This is just a first step and adds a new role called
kubernetes_vanilla that executes distro specific tasks.
Currently, only the kubelet.service file is configured via the new
mechanism.

Signed-off-by: Thomas Bechtold <tbechtold@suse.com>