ansible/roles/kubernetes_vanilla: Rename Suse to common & drop main
There is no SUSE specifics in the role tasks. Rename that to
common.yaml and run it on all hosts. Also drop the default main.yaml
which is no longer needed.
There will be another tasks file which contains master specific code
that will be only executed on the k8s master nodes.

Signed-off-by: Thomas Bechtold <tbechtold@suse.com>