kubernetes: Drop workspace from KubernetesBase constructor
We already pass an Hardware instance to the kubernetes constructor and
Hardware contains already a workspace instance.
There is no need to have to workspace instances. In fact this is
confusing and given that we pass the workspace_dir now to ansible
playbooks (see commit 6e74f4b5169c4bf), this is even more confusing
given that it's not clear which workspace instance this might be.

Note: We keep the shortcut to access the workspace instance from
kubernetes. So:

kubernetes.workspace.foo() is still possible (instead of
kubernetes.hardware.workspace.foo())

Signed-off-by: Thomas Bechtold <tbechtold@suse.com>