rook: Drop workspace from RookBase constructor
We already pass an Kubernetes instance to the rook constructor and
kubernets.hardware contains already a workspace instance.
There is no need to have two workspace instances. In fact this is
confusing and given that we pass the workspace_dir now to ansible
playbooks (see commit 6e74f4b5169c4bf), this is even more confusing
given that it's not clear which workspace instance this might be.

Signed-off-by: Thomas Bechtold <tbechtold@suse.com>