hardware: Add abstractmethod node_create()
This is needed to be able to create new nodes within tests. Currently
there is only node_add() which expects a NodeBase object as
parameter. But this NodeBase object needs to be created somehow and
the creation is implementation specific (eg. different between libvirt
and libcloud-openstack).
Adding a node_create() method that returns a NodeBase object will
solve that problem.

Signed-off-by: Thomas Bechtold <tbechtold@suse.com>