workspace: Move basic ansible inventory vars to workspace
Instead of having the basic ansible variables (eg. private key, host
key checking, workspace dir, ...) in the hardware base class, move
theses to the workspace class.
That's logical given that the variables used there are defined in the
Workspace class anyway.
This also simplifies the variables that need to be passed into the
ansible executor methods.

Signed-off-by: Thomas Bechtold <tbechtold@suse.com>