Add a function to add disks
Rook does not deploy properly without dedicated data disk (i.e
using a folder as a fake 'data holder'). This is at least the case
for SES rook.

This PR add a function to add disks to node. The size defaults to
10G as of now but this should be improved by allowing both the
disk number and size to be specified as global test variables.