hardware/hardware_base: Add basic node-base ansible structure
This is the beginning of moving to pure ansible code (instead of using
the unstable ansible API from python).
There will be a new "node-base" role that is responsible for setting
up the nodes. Distro specific code can be executed there.

As a first step, move the "wait for nodes" code to the new node-base
role. Also move the setup for "max open files" and "swappiness" and
fix it so it survives a node reboot.

Signed-off-by: Thomas Bechtold <tbechtold@suse.com>