Drop ansible-helper and move methods to hardware base
There is no need to have the ansible-helper code anymore. The
AnsibleRunner class took a workspace instance and a list of nodes -
both are already available in the HardwareBase class.
So having the code to trigger a ansible-playbook run in the Hardware
base class makes sense.
This also solves the problem to recreate the inventory structure when
a new node was added or removed. This is now simply done in the
"node_add()" and "node_remove()" methods. So the inventory/ directory
is always up-to-date.

Signed-off-by: Thomas Bechtold <tbechtold@suse.com>