hardware/openstack_libcloud: Automatically create net/subnet/router
When creating a new test environment, automatically create a network,
a subnet and a router which can be used by the nodes.
This replaces the manual steps that were documented. It also removes
the OS_INTERNAL_NETWORK config option which is no longer needed.
Note: We need to set "api_version=2.2" for the OpenStack connection to
use the newer OpenStack NodeDriver class within libcloud.

Signed-off-by: Thomas Bechtold <tbechtold@suse.com>