Automatically choose network subnet for libvirt
The libvirt driver creates a network for virtual machines. However if you
are building more than one instance of hardware (for example, when running
lots of tests in parallel), the network creation fails because it already
exists.

This patch breaks a network range into subnets for each test run. It tries
a subnet until it finds an available one.