lib/hardware: Add openstacksdk base OpenStack hardware provider
Add another hardware provider that is based on the official OpenStack
SDK.
That is an alternative to the available apache-libcloud provider. The
new provider can be used by setting
ROOKCHECK_HARDWARE_PROVIDER=OPENSTACKSDK. There are a couple of
advantages over apache-libcloud:

- hopefully better tested and widely used (we ran into a couple of
  apache-libcloud bugs recently)
- support for most/all OpenStack features - including
  LoadBalancers (which are not supported in apache-libcloud)
- support for clouds.yaml files (via OS_CLOUD)

Given that we have anyway another abstraction layer on top of
apache-libcloud (for the libvirt provider), there is no real need to
use the apache-libcloud abstraction.

Signed-off-by: Thomas Bechtold <tbechtold@suse.com>