Set hostname explicitly
On openstack and libvirt the hostnames were automatically set (probably
by cloudinit). The opensuse images in aws do not seem to set the hostname
(or at least boot before we can tag it). Either way it is better to
explicitly set it to the node name that we/ansible expects.