lib/hardware: Improve disk_create method and its usage
Make it clear that the disk_create() method expects an int for the
disk capacity in GB.
This also fixes the disk deployment for libvirt where
test_add_storage() test failed because the rook-discover pod didn't
use the newly created disk (due to "Insufficient space (\u003c5GB)").
The extra qcow2 disk that was created for libvirt had a size of 512
byte.

Signed-off-by: Thomas Bechtold <tbechtold@suse.com>