Add option to leave cluster available after run
By default the cluster is destroyed at each run. This adds a new option
(_tear_down_cluster) which can be set to False to keep the cluster around.

This also tidies up the removal of AWS nodes which can use the super
class's method to destroy them instead of repeating the code.