Do not log traceback for failed pod log collection
We crudely iterate over the pods to grab their logs and other details.
However there can be a race where the pod no longer exists before we
grab their logs. This is a common failure and not an important one
so instead of logging the traceback, just add a small warning.

The details of the missed pods will be captured by the preceeding
methods.